PP-7786: restrict switch to test payouts page to test stripe
Only show the "Switch to test accounts" link if the user viewing the
page has access to a test Stripe accounts.

Until there is analysis to review if we should be showing all of the
tools to all users this is to avoid live users getting confused as to
why their test "sandbox" accounts aren't showing up on this view.