PP-7733 Remove unnecessary middleware
Remove the getAccount, hasServices and resolveService middleware on all
routes that no longer require it.

Replace usages of the `enforceUserAuthenticated` middleware with the new
`is-user-authorised` middleware which does all the same checks.

Change tests that expect a redirect to `/otp-login` if the user has not
completed second factor authentication. This was never actually the
behaviour, as there was always a check on the session version first
which is updated after second factor authentication was completed, and
if the session version didn't match we redirected to `/login`. The
new `is-user-authorised` middleware now just always directs to `/login`
so this ambiguity is removed without changing actual behaviour.