PP-7715 Fix redirect after refund
Fix the redirect after making a refund if the user has come from the
"Transactions for all services" page, rather than the transactions list
for a particular account.

Do this by adding a hidden input to the refund form to indicate that we
have come from the all services transactions page which is passed along
with the redirect.

Store the filters for the all services transaction search on the session
as we do for the account transaction search so we can construct the back
link from these, even after redirecting following a refund. Also include
the status filter (live/test) to ensure we get back to the correct
search.