PP-7735 Use service router for service routes
Add a new Router for service routes, following the pattern we use for
account routes. On this router, use the new
`get-service-and-gateway-account` middleware, which replaces
`resolve-service`, and the `is-user-authorised` middleware.

In `paths.js` group the routes under a `service` object, as we do for
account routes and update all usages.

Update the `permission` middleware to no longer call `resolve-service`
and delete the now unused `resolve-service` and `has-services`
middleware.