PP-7735 Make service URLs uniform
Add new URLs for the organisation details page that use the pattern
`/service/:externalServiceId` so that we can use a Router for all
service routes.

Also change the URL for view/edit team members to be
`/service/:externalServiceId/team-members` rather than
`/service/:externalServiceId/` as this isn't really a root URL for a
service.

Keep routes for the old URLs which will be removed after this change
has been deployed. This is to avoid breaking user journeys during
deployment.