PP-7735 Inline userIsAuthorised middleware
Remove the array of `authenticatedPaths` to apply this middleware to
and inline the middleware per route. This should make it clearer what
middleware is being called on the routes and should make it harder to
make mistakes when adding a new route.

The unit test for `routes.test.js` ensures that all routes using the
main Express router call this middleware unless they are specifically
excluded which adds protection against coding mistakes.