PP-7734 Remove get-gateway-account middleware
The only thing this middleware was still used for was setting the
gateway account cookie to the first gateway account found on login if
it wasn't already set.

The cookie is still used for upgrading old account URLs without the
external ID to the new URLs based on the account in the cookie. We will
now redirect to the "My services" page if no cookie is set.