PP-7570 Use AsyncLocalStorage to provide shared logging context
Add middleware to open an AsyncLocalStorage store on a request. Add the
correlation ID when the store is initialised.

Add a format function to the logger that pulls fields that have been
added to the store and adds them to the log line.

Add middleware to add user info to the log context when it is available.
Add account and service info to the log context in the middleware that
resolves them.

Re-order the server middleware so that the auth middleware is run after
the global middleware that initialises logging amongst other things.