BAU: inlude policy pages in non-service/account templates
Policy document pages were being shown as account pages before the
refactored URL based account routing. This behaviour is not desired as
the it would have been the previous account the user had selected, which
will never be directly relevant to the policy document selected.

Include the template path in the non-service/ account pages until the
template inheritance method is reviewed.