BAU: fix pact unit test not waiting on before
- return the result of pact provider setup to make sure that it's
available when the test runs
- Also simplify assertion for flakey worldpay flex pact

I'm not familiar with the chai `should` syntax but this unit test seems
to fail every so often for no reason.

Use async/ await syntax to make it clearer what's going on and review
behaviour to make sure it will only return once the request and
assertion have completed.