PP-8218: split out stripe pages to include switch psp routes
Namespace Stripe pages by `/your-psp/` and `/switch-psp/` to line them
up with other credential specific pages.

The credential ID namespaced is not currently used as the Stripe
progress flags are set at the account level but it should be simple to
use the credential details when they move over to gateway account
credentials.