PP-8728 Service reg: no OTP pages without password on invite
In the service registration flow, when loading the page where
the user enters the OTP or the page where the user can choose to
resend the OTP, retrieve the invite (which will check it’s still
valid) and make sure it has the flag indicating the password has
been set.

This prepares us for a world where the invite may not have a
password set when it is first created.