PP-8750 Simplify credentials controller and views
Notification credentials are only a concept that apply to smartpay,
but in the controller we have the concept of displaying the credentials
view in "notification credentials mode".

Simplify this by separating out the notifications credentials page into
a separate view, and modify the controller so there isn't shared code
for editing the account credentials and the notification credentials.