PP-8715 Better link for Microsoft Authenticator
Update the Microsoft Authenticator app link (shown to a user on the
page that lets them choose their sign-in method) so that it goes
to a friendly product page with obvious calls to action for
installing the app rather than a wordy support page that explains
how to download and install it.

This is still not perfect as the new page links to support docs
that assume the user wants to use the app with a Microsoft account
but the old page also suffered from this (and the Google
Authenticator page we direct users to is even worse).