PP-8750 Remove client side validation from payment links
Reproduce all client-side validation on the server side, and remove
the client side validation from the payment links pages.

Removing client-side validation as it is against design system guidance,
makes it difficult to test validation, and is not consistent with the
server side validation.

For checking whether fields contain characters not allowed by NAXSI,
don't check for a ";", as we have this on the allow list for most routes