PP-8750 Remove client-side valdition for demo payment
Reproduce client-side validation on the server for the edit demo payment
details journey.

Removing client-side validation as it is against design system guidance,
makes it difficult to test validation, and is not consistent with the
server side validation.

Restructure the controllers to split out the edit controllers so the
code is easier to understand.

Replace the functional tests with Cypress tests, and add unit tests for
the edit amount controller to test all possible validation errors.