PP-8750 Remove more client-side validation
Removing client-side validation as it is against design system guidance,
makes it difficult to test validation, and is not consistent with the
server side validation.

For most of this client-side validation, the same validation already
exists on the server. Add in server-side validation that was missing
for email-notifications.controller.

Amend the wording for the login screen errors to make them more
consistent with design system guidance.