PP-8830 Reuse same controller for both organisation details pages
There are two pages where users can enter organisation details:
- A page accessible from "My services"
- As part of the request to go live journey

The main difference between these pages is that for the request to go
live journey, the organisation name is collected on a separate page.

The other page has been left behind when content changes have been made,
and having them separate allows them to keep getting out of sync.

Combine the pages to use the same controllers and view, but for the
page that is accessed from "My services" also collect the organisation
name.

Did some additional tidy up - renamed some code references to merchant
details to organisation details to reflect the name of the page. Removed
the functional tests and replaced them with Cypress tests.