PP-8928 Fix link in error summary
For radio buttons, there is no ID for the containing fieldset, so error
summary messages need to link to the first radio button item. Fix the
ID for the error on the VAT number page so it links to the first radio
button item, which has a custom ID.

Also, use the same inline error message as the error message in the
error summary to be consistent with the design system and other pages.