PP-8728 WIP - Make set phone number and password page reusable
Move the page for setting the phone number and password when a
user signs up by accept an invitation to a common location so
that it can be reused in the flow when a user self-initiates their
account creation.

This is a work in progress. Functionally it works but the tests
need progress. The idea was to use a Cypress test and unit tests
to replace some of the existing integration tests.