pp-8989 allow non GB prefixed VAT numbers
Previously, we thought VAT numbers had to be prefixed with GB but that is no longer the case. Wrote a new regex function to correctly validate all types of UK VAT numbers with or without the GB prefix. Updated the view to follow HMRC suggested guidelines around taking VAT numbers from customers.