PP-9052 Fix name service during account creation bug
Ensure that the service we are setting the name for as the last step of
a user signing up for an account is the test service created for them
during the sign-up process.

There was an issue where if the user started registering for an account
and then accepted an invite to join a service before completing it, they
could end up changing the name of the service they were invited to by
mistake, as our code was only expecting them to have one service at this
point.

To avoid this, check for a service with the default name, which is
always "System Generated", and rename that service. If the user does not
have a service with the default name, show an error page.

If the user tries to access the page for setting the service name when
they do not have a service with the default name, redirect to the "My
services" page.

Extract the default service name, which is "System Generated" into a
constant referenced from all places we use it in application code.