PP-9052 Fix service name validation
Service name validation wasn't working for naming a service during
account registration. If a service name that is too long is submitted,
the input was just cleared and no error message was shown.

While fixing this, also improved the error messages for service name
validation on all pages. Previously, if the name was too long the
message was "The text is too long". Improve this to state how the
character limit. Also move to using the design system recommendations
for displaying the errors in the error summary.