PP-9148 Cache Cypress with version in key
In our workflow for running tests on PRs, we cache Cypress so that it
doesn't need to be downloaded and installed for each test run, as this
takes a number of minutes. This cache is persisted between action runs.

We were caching with the key `Linux-cypress` not taking into account
versioning. This means that if we tried to update our Cypress version,
we would probably end up with multiple Cypress versions in the same
cache due to restoring the existing version from the cache and then
installing the new version when we run `npm install`.

To ensure that we only cache and restore the required version, retrieve
the Cypress version number from the package.json file and append this
to the cache key.