PP-9460 Combine amount and edit amount templates
The template for specifying the amount for a new payment amount and
editing the amount for an existing payment link are almost identical.
Combine these and use the template property 'isEditing' to decide which
content to display that differers depending on the context.

Removed some template properties from the controllers which were
unused.