PP-9460 Combine reference and edit reference templates
The template for giving details about the reference for a new payment
link and editing the reference details for an existing payment link are
almost identical.
Combine these and use the template property 'isEditing' to decide which
content to display that differs depending on the context.

Rename the session variables to be consistent with the edit journey to
increase the scope of unifying these journeys in the future. Ensure that
in-progress payment link journeys will continue to work with the old
session variable names, and we can remove the temporary code to allow
this a while after the change has been deployed.