PP-9460 Collect amount hint for payment links
When the option that the user can pick the amount is clicked, display
a text input that allows them to specify a hint to be shown on the
amount page.

Send this hint when creating the product.

Hide this behind a new feature flag, PAYMENT_LINKS_FUTURE_ENABLED so it
is not shown unless this feature flag is enabled. For editing a payment
link show the input either if the environment variable is enabled, or
the `new_payment_link_journey_enabled` flag is true for the product.

On the review and edit pages, make all text in the value columns other
than the hints use black font rather than grey font. For example, when
the value is "User can choose" for the amount.