BAU — Remove LastPass Authenticator link and add FreeOTP link
In our list of suggested authenticator apps, remove LastPass
Authenticator because the page we link to —
https://lastpass.com/auth/ — now redirects to
https://www.lastpass.com/solutions/authentication and extols the
virtues of ‘Customizable authentication methods with admin
oversight’ and tries to persuade me to sign up for LastPass
Business. There is no link to download LastPass Authenticator any
more (the app does still seem to be available in the iOS App
Store and Google Play but the only page on the LastPass site with
download links seems to be a support page at
https://support.lastpass.com/help/enroll-the-lastpass-authenticator-app
where the download links are buried in the middle of a bunch of
steps for enabling 2FA on a LastPass vault, which is not user-
friendly).

Replace it with a link to FreeOTP, which is a simple authenticator
app published by Red Hat.