PP-9587: Display banner on dashboard for disabled accounts
If gateway account has the `disabled` flag set to true, display a
banner on the dashboard that indicates the account is disabled. Do not
display any of the other possible banners if this is displayed.

Add a Cypress test for this. The Pact with connector is updated to
expect the `disabled` flag, as we've updated the fixture used to
generate the Pact.

Co-authored-by: Alan Carter <alan.carter@digital.cabinet-office.gov.uk>