PP-9846: add agreement id to advanced filter
Transactions can be filtered by their agreement identifier, if
they've been taken on behalf of a recurring card payment agreement.

This adds an additional filter to the "Advanced filter" element, this
has the benefit of not confusing users which this filter has no
relevance for however where and how this filter fits in the with the
larger transactions page should be considered in future design.