PP-9846: only show recurring filters to appropriate accounts
For the time being the admin tool only shows the agreement tab to
services with the `recurring_enabled` flag set to true (as managed by
Toolbox). Reflect this with the transactions filter until a decision is
made on if this should be shown to all users.

Update the content to read "Search by agreement" followed by the hint to
"Enter agreement identifier".