PP-6213 Make extremely long payment references not overflow
Make extremely long payment references with no spaces not overflow
on both the transactions list page and the transaction details
page.

This has been a problem since forever but it’s more pressing
now that we allow payment link references (which may be entered
by paying users without oversight from the partner service that’s
going to see them in the admin tool) to be up to 255 characters
in length (up from 50).

We use table-layout: fixed and overflow-wrap: break-word to force
jumbo references to wrap. For compatibility with browsers that do
not support overflow-wrap (mostly IE11), we also use the legacy
word-wrap property.

On the transaction details page, this will also stop extremely
long email addresses and cardholder names from causing mischief.