PP-10047 Add page to provide phone number
When a user has registered their account using an authenticator app to
get security codes, which is currently being implemented, they won't be
asked for a phone number.

If they then go to change their sign-in method to "Text message" from
their profile settings, add a page that asks for their phone number.
This page is shown after selecting the new sign-in method, if "Text
message" is selected and the user does not have a phone number set.

Upon submitting the new phone number, make a PATCH request to adminusers
to update the phone number on the user, then make requests to provision
a new provisional OTP key and to send a text message with a security
code.