PP-10240 Remove the "old" client error handler
There is currently the concept of an "old" and a "new" error handler.
The "old" error handler puts the entire response body in the message
field of the "error" object returned by the client.

There was only one place in the code where fields in the response body
was accessed - when attempting to complete a registration. Here we were
using the backend error message to display on the error page. We
shouldn't do this, so instead just display a generic error page.

Remove the "old" error handler so that the clients return an error
with a uniform signature when a request returns a non-success response.
This error contains the error message string extracted from the HTTP
response body.