PP-10240 Introduce RESTClientError type
Introduce a new type which extends Error for representing errors thrown
by the REST client. This makes the API of these errors clearer and
makes it possible to handle these in a specific way.

Handle uncaught RESTClientErrors in the error handler middleware in
order to log more detail about the error.

Rewrite `wrapper.test.js` to be easier to understand.