PP-10047 Render error page for errors updating 2FA method
Previously, if there were unexpected errors returned by adminusers in
the controllers for updating the second factor method, we redirected the
user back to the my services page and showed an error notification box
at the top of the page.

This is not a pattern we use elsewhere anymore, instead we render an
error page. Make these controllers consistent with behaviour elsewhere
and render an error page.