PP-10047 Allow user to change number when resending OTP code
On the page to enter the verification code when updating the sign-in
method to SMS, there is a link that when clicked will send another code
to the configured phone number.

Change this link so that instead of sending a code to the same phone
number, the user is taken to another page where they are shown the phone
number they have configured and are given the opportunity to update it
and then send a new code.

This matches the user journey for configuring second factor
authentication during registration.

Also update the text of the link to resend the code to the new text that
was agreed upon when reviewing the content.