PP-10276: remove correlation ids from fixtures
Clean up any remaining `correlationId` values that were in test
fixtures, now that the correlation id middleware has been removed from
the stack these will no longer be on the request (in favour of using
async hooks for logging and client requests).

Additionally removes another helper and tests that seemed to only be providing the
`CORRELATION_ID` header which is already exported by a config file.