PP-10037: fix google analytics redact url
We were previously setting the 'path' attribute on Google Analytics,
this is submitted with a `dp=` attribute and had appropriately redacted
values, however the full location is also sent through `dl=` which
didn't included the full url.

These fields are documented
https://developers.google.com/analytics/devguides/collection/analyticsjs/field-reference#location.

The "Document Path" which we were previously setting seems to be set up
for "specify virtual page paths.", as we have no dynamic or client side
managed navigation, "Document Location" feels like the appropriate place
to redact and send this information.

Set the 'location' attribute instead to ensure everything that is sent
to GA is appropriately redacted.

Any pages which may be impacted have been filtered out of GA until this
ticket had been addressed.