PP-10116 update secrets baseline, pre-commit conf
- update secrets baseline
- update pre-commit conf
- add detect secrets GHA

PP-10244 Render "My services" page for no services

A user can be a member of no services if they have been removed from all
services they were a member of. Currently the controller for showing
the "My services" page throws an error if the user is a member of no
services and an error page is rendered.

Fix this so that the controller does not throw an error, and the page is
rendered. The rendered page does not include the "Reports" section and
has an additional paragraph instructing the user on what to do next.

Add some additional Cypress tests to check we're displaying the
"My services" page correctly in different scenarios.