PP-10244 Handle service not created during registration
We are going to remove creating a service as part of a self-signup
registration.

The service is currently created by adminusers when a request is made to
the `/v1/api/invites/{code}/complete` endpoint, which also creates the
user account. Selfservice then redirects the user to a page to name the
created service.

We are going to stop creating a test service in adminusers when
`/v1/api/invites/{code}/complete` is called. When we are no longer
creating a service, the response from this endpoint will no longer
include a `service_external_id` field.

Prepare for a service no longer being created by handling the case where
`/v1/api/invites/{code}/complete` does not include a
`service_external_id`. In this case, do not attempt to create a gateway
account in connector. And instead of redirecting the user to the page to
set the service name after logging them in, redirect the user to the
"My services" page.