PP-10055 Add pact test for GET invite
Amend get-invite.pact.test.js which previously wouldn't be verified
against adminusers as it had consumer `selfservice-to-be` so that it
will successfully run against adminusers.

Remove tests that check error responses, as we handle these in a
standard way.

Fix the invite fixture so that it will not insert properties that are
not defined in the fixture  - using `lodash.defaults` was using any
fields passed in in the `opts` object even if they were not defined in
the fixture. Remove the fields from the fixture that selfservice does
not consume.