PP-10055 Add route to submit password page
Add a POST `/register/password` route that is POSTed to when the
password page is submitted. This does validation of the
"Enter a password" and "Re-type your password" fields.

If the password is valid, a PATCH request is made to adminusers to
update the password and we redirect to the next page in the journey.