PP-10052 Strip spaces and dashes from OTP codes
In the design system page for the pattern to enter OTP codes, it says to

"Let the user enter the code in whatever format is familiar to them.
Allow additional spaces, hyphens and dashes."

The reasoning for this is that speech-to-text programs may cause spaces
to be entered in the code, or authenticator apps might provide the code
in a different format using spaces or dashes.

For all places where we ask the user to enter an OTP code, remove spaces
and dashes before validating and submitting the code.