PP-10052 Use constants for invalid OTP error messages
To ensure that we are using the same error message in all places that
validate OTP codes, move the validation messages we display when
adminusers returns a 401 response when verifying the code to shared
constants.

Also, display an improved error message if the OTP method is to use an
authenticator app. This message is taken from the GOV.UK Sign-in
login/registration pages.