PP-10052 Improve OTP code validation for login
Currently when an OTP code is entered during the user login journey, we
send it directly to adminusers without doing any validation. This means
that we do not display any specific validation messages if the the
code is the wrong length etc.

When the code is entered, the validation and verification is done by the
`localStrategy2Fa` in `auth.service.js`. This is registered with
passport, which is the library we use to manage user sessions. This is
called by `login-user-otp.controller.js` as part of the login journey.

Add some missing tests for the `localStrategy2Fa` method, and some
Cypress tests to check validation messages are displayed correctly on
the OTP code page.

Fix the error summary in the template to use the correct header -
"There is a problem". While doing this, switch to using the Nunjucks
macro rather than HTML.