PP-10057 Add middleware to check registration cookie is present
To avoid checking in every controller for the registration pages
whether the session cookie for registration is present, add middleware
to check it is present, else call next with an error so that an
appropriate error page is displayed to the user.

Add a constant for the session cookie name to avoid hardcoding it in
lots of places.