PP-10052 Send second factor method when completing invite
After successfully verifying the OTP code on the page to configure an
authenticator app, we make a request to
/v1/api/invites/${inviteCode}/complete on adminusers to complete the
invite and create the user. Make it so that we send
"second_factor": "APP" in the request body so that the user is created
with "APP" set as their second factor method for logging in.