PP-10052: Fix: pass second factor method through for invite client
Now that second factor method will only be included in the invite
complete request if it is passed into the client method, all consuming
code should set the value.

Propegate the value through the client method but also set a default to
SMS which is currently used for all invites.